third_party/WebKit/LayoutTests/fast/dom/Document/scripts-non-html.html

<!DOCTYPE html>
<html>
<head>
<link rel="help" href="http://www.whatwg.org/specs/web-apps/current-work/multipage/dom.html#dom-document-scripts"/>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Make sure the script attribute is exposed on non-HTML documents.");

var xmlDocument = document.implementation.createDocument("", "", null);
shouldBe("xmlDocument.__proto__", "XMLDocument.prototype");
shouldBe("xmlDocument.__proto__.__proto__", "Document.prototype");

shouldBe("xmlDocument.scripts.__proto__", "HTMLCollection.prototype");
shouldBe("xmlDocument.scripts.length", "0");
</script>
</body>
</html>
