third_party/blink/web_tests/fast/dom/Document/scripts-non-html-expected.txt

Make sure the script attribute is exposed on non-HTML documents.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xmlDocument.__proto__ is XMLDocument.prototype
PASS xmlDocument.__proto__.__proto__ is Document.prototype
PASS xmlDocument.scripts.__proto__ is HTMLCollection.prototype
PASS xmlDocument.scripts.length is 0
PASS successfullyParsed is true

TEST COMPLETE
